+++ samples/subsys/mgmt/mcumgr/smp_svr/overlay-bt-tiny.conf
@@ -1,0 +1,18 @@
+### This configuration enables fewer settings than the default, making it
+### suitable for a device with 32kB of RAM and 72kB image slots (e.g.,
+### nRF51).
+
+# Allow for large Bluetooth data packets.
+CONFIG_BT_L2CAP_TX_MTU=260
+CONFIG_BT_RX_BUF_LEN=260
+
+# Enable the Bluetooth (unauthenticated) and UART mcumgr transports.
+CONFIG_MCUMGR_SMP_BT=y
+CONFIG_MCUMGR_SMP_BT_AUTHEN=n
+CONFIG_MCUMGR_SMP_UART=y
+
+# Disable Bluetooth unused features
+CONFIG_BT_GATT_READ_MULTIPLE=n
+
+# Disable statistics and statistic names.
+CONFIG_STATS_NAMES=n
