+++ samples/subsys/mgmt/mcumgr/smp_svr/prj.conf
@@ -7,22 +7,8 @@ CONFIG_SYSTEM_WORKQUEUE_STACK_SIZE=2048
 # Ensure an MCUboot-compatible binary is generated.
 CONFIG_BOOTLOADER_MCUBOOT=y
 
-# Allow for large Bluetooth data packets.
-CONFIG_BT_L2CAP_TX_MTU=260
-CONFIG_BT_RX_BUF_LEN=260
-
-# Enable the Bluetooth (unauthenticated) and shell mcumgr transports.
-CONFIG_MCUMGR_SMP_BT=y
-CONFIG_MCUMGR_SMP_BT_AUTHEN=n
-CONFIG_MCUMGR_SMP_SHELL=y
-#CONFIG_MCUMGR_SMP_UART=y
-
 # Enable flash operations.
 CONFIG_FLASH=y
-
-# Enable the LittleFS file system.
-CONFIG_FILE_SYSTEM=y
-CONFIG_FILE_SYSTEM_LITTLEFS=y
 
 # Required by the `taskstat` command.
 CONFIG_THREAD_MONITOR=y
@@ -31,8 +17,10 @@ CONFIG_THREAD_MONITOR=y
 CONFIG_STATS=y
 CONFIG_STATS_NAMES=y
 
-# Enable all core commands.
-CONFIG_MCUMGR_CMD_FS_MGMT=y
+# Enable most core commands.
 CONFIG_MCUMGR_CMD_IMG_MGMT=y
 CONFIG_MCUMGR_CMD_OS_MGMT=y
 CONFIG_MCUMGR_CMD_STAT_MGMT=y
+
+# Enable logging
+CONFIG_LOG=y
