+++ samples/subsys/mgmt/mcumgr/smp_svr/src/bluetooth.c
@@ -1,0 +1,89 @@
+/*
+ * Copyright (c) 2012-2014 Wind River Systems, Inc.
+ * Copyright (c) 2020 Prevas A/S
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+#include <bluetooth/bluetooth.h>
+#include <bluetooth/conn.h>
+#include <bluetooth/gatt.h>
+#include <mgmt/smp_bt.h>
+
+#define LOG_LEVEL LOG_LEVEL_DBG
+#include <logging/log.h>
+LOG_MODULE_REGISTER(smp_bt_sample);
+
+static struct k_work advertise_work;
+
+static const struct bt_data ad[] = {
+	BT_DATA_BYTES(BT_DATA_FLAGS, (BT_LE_AD_GENERAL | BT_LE_AD_NO_BREDR)),
+	BT_DATA_BYTES(BT_DATA_UUID128_ALL,
+		      0x84, 0xaa, 0x60, 0x74, 0x52, 0x8a, 0x8b, 0x86,
+		      0xd3, 0x4c, 0xb7, 0x1d, 0x1d, 0xdc, 0x53, 0x8d),
+};
+
+static void advertise(struct k_work *work)
+{
+	int rc;
+
+	bt_le_adv_stop();
+
+	rc = bt_le_adv_start(BT_LE_ADV_CONN_NAME, ad, ARRAY_SIZE(ad), NULL, 0);
+	if (rc) {
+		LOG_ERR("Advertising failed to start (rc %d)", rc);
+		return;
+	}
+
+	LOG_INF("Advertising successfully started");
+}
+
+static void connected(struct bt_conn *conn, u8_t err)
+{
+	if (err) {
+		LOG_ERR("Connection failed (err 0x%02x)", err);
+	} else {
+		LOG_INF("Connected");
+	}
+}
+
+static void disconnected(struct bt_conn *conn, u8_t reason)
+{
+	LOG_INF("Disconnected (reason 0x%02x)", reason);
+	k_work_submit(&advertise_work);
+}
+
+static struct bt_conn_cb conn_callbacks = {
+	.connected = connected,
+	.disconnected = disconnected,
+};
+
+static void bt_ready(int err)
+{
+	if (err) {
+		LOG_ERR("Bluetooth init failed (err %d)", err);
+		return;
+	}
+
+	LOG_INF("Bluetooth initialized");
+
+	k_work_submit(&advertise_work);
+}
+
+void start_smp_bluetooth(void)
+{
+	k_work_init(&advertise_work, advertise);
+
+	/* Enable Bluetooth. */
+	int rc = bt_enable(bt_ready);
+
+	if (rc != 0) {
+		LOG_ERR("Bluetooth init failed (err %d)", rc);
+		return;
+	}
+
+	bt_conn_cb_register(&conn_callbacks);
+
+	/* Initialize the Bluetooth mcumgr transport. */
+	smp_bt_register();
+}
