+++ samples/subsys/mgmt/mcumgr/smp_svr/src/common.h
@@ -1,0 +1,8 @@
+/*
+ * Copyright (c) 2020 Prevas A/S
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+void start_smp_bluetooth(void);
+void start_smp_udp(void);
