+++ samples/subsys/mgmt/mcumgr/smp_svr/src/main.c
@@ -1,15 +1,12 @@
 /*
  * Copyright (c) 2012-2014 Wind River Systems, Inc.
+ * Copyright (c) 2020 Prevas A/S
  *
  * SPDX-License-Identifier: Apache-2.0
  */
 
-#include <assert.h>
 #include <zephyr.h>
-#include <string.h>
-#include <stdlib.h>
 #include <stats/stats.h>
-#include <mgmt/buf.h>
 
 #ifdef CONFIG_MCUMGR_CMD_FS_MGMT
 #include <device.h>
@@ -27,12 +24,11 @@
 #include "stat_mgmt/stat_mgmt.h"
 #endif
 
-#ifdef CONFIG_MCUMGR_SMP_BT
-#include <bluetooth/bluetooth.h>
-#include <bluetooth/conn.h>
-#include <bluetooth/gatt.h>
-#include <mgmt/smp_bt.h>
-#endif
+#define LOG_LEVEL LOG_LEVEL_DBG
+#include <logging/log.h>
+LOG_MODULE_REGISTER(smp_sample);
+
+#include "common.h"
 
 /* Define an example stats group; approximates seconds since boot. */
 STATS_SECT_START(smp_svr_stats)
@@ -57,76 +53,20 @@ static struct fs_mount_t littlefs_mnt = {
 };
 #endif
 
-#ifdef CONFIG_MCUMGR_SMP_BT
-static struct k_work advertise_work;
-
-static const struct bt_data ad[] = {
-	BT_DATA_BYTES(BT_DATA_FLAGS, (BT_LE_AD_GENERAL | BT_LE_AD_NO_BREDR)),
-	BT_DATA_BYTES(BT_DATA_UUID128_ALL,
-		      0x84, 0xaa, 0x60, 0x74, 0x52, 0x8a, 0x8b, 0x86,
-		      0xd3, 0x4c, 0xb7, 0x1d, 0x1d, 0xdc, 0x53, 0x8d),
-};
-
-static void advertise(struct k_work *work)
-{
-	int rc;
-
-	bt_le_adv_stop();
-
-	rc = bt_le_adv_start(BT_LE_ADV_CONN_NAME, ad, ARRAY_SIZE(ad), NULL, 0);
-	if (rc) {
-		printk("Advertising failed to start (rc %d)\n", rc);
-		return;
-	}
-
-	printk("Advertising successfully started\n");
-}
-
-static void connected(struct bt_conn *conn, u8_t err)
-{
-	if (err) {
-		printk("Connection failed (err 0x%02x)\n", err);
-	} else {
-		printk("Connected\n");
-	}
-}
-
-static void disconnected(struct bt_conn *conn, u8_t reason)
-{
-	printk("Disconnected (reason 0x%02x)\n", reason);
-	k_work_submit(&advertise_work);
-}
-
-static struct bt_conn_cb conn_callbacks = {
-	.connected = connected,
-	.disconnected = disconnected,
-};
-
-static void bt_ready(int err)
-{
-	if (err) {
-		printk("Bluetooth init failed (err %d)\n", err);
-		return;
-	}
-
-	printk("Bluetooth initialized\n");
-
-	k_work_submit(&advertise_work);
-}
-#endif
-
 void main(void)
 {
-	int rc;
+	int rc = STATS_INIT_AND_REG(smp_svr_stats, STATS_SIZE_32,
+				    "smp_svr_stats");
 
-	rc = STATS_INIT_AND_REG(smp_svr_stats, STATS_SIZE_32, "smp_svr_stats");
-	assert(rc == 0);
+	if (rc < 0) {
+		LOG_ERR("Error initializing stats system [%d]", rc);
+}
 
 	/* Register the built-in mcumgr command handlers. */
 #ifdef CONFIG_MCUMGR_CMD_FS_MGMT
 	rc = fs_mount(&littlefs_mnt);
 	if (rc < 0) {
-		printk("Error mounting littlefs [%d]\n", rc);
+		LOG_ERR("Error mounting littlefs [%d]", rc);
 	}
 
 	fs_mgmt_register_group();
@@ -140,21 +80,10 @@ void main(void)
 #ifdef CONFIG_MCUMGR_CMD_STAT_MGMT
 	stat_mgmt_register_group();
 #endif
-
 #ifdef CONFIG_MCUMGR_SMP_BT
-	k_work_init(&advertise_work, advertise);
-
-	/* Enable Bluetooth. */
-	rc = bt_enable(bt_ready);
-	if (rc != 0) {
-		printk("Bluetooth init failed (err %d)\n", rc);
-		return;
-	}
-	bt_conn_cb_register(&conn_callbacks);
-
-	/* Initialize the Bluetooth mcumgr transport. */
-	smp_bt_register();
+	start_smp_bluetooth();
 #endif
+
 
 	/* The system work queue handles all incoming mcumgr requests.  Let the
 	 * main thread idle while the mcumgr server runs.
